move progressBar
