+++ platform-sponge-api7/src/main/kotlin/taboolib/platform/util/TextUtil.kt
@@ -2,7 +2,6 @@
 
 package taboolib.platform.util
 
-import com.google.common.base.Strings
 import org.spongepowered.api.text.Text
 import org.spongepowered.api.text.serializer.TextSerializers
 import taboolib.common.Isolated
@@ -21,25 +20,4 @@ fun String.toTextWithFormattingCode(): Text {
  */
 fun Text.toPlainWithFormattingCode(): String {
     return TextSerializers.FORMATTING_CODE.serialize(this)
-}
-
-/**
- * 百分比状态
- * @param current 当前值
- * @param max 最大值
- * @param totalBars 总值
- * @param symbol 标识符
- * @param completedColor 完成的颜色
- * @param notCompletedColor 未完成的颜色
- * @return
- */
-fun progressBar(
-    current: Int, max: Int, totalBars: Int, symbol:
-    String, completedColor: String,
-    notCompletedColor: String): String {
-    val now = if (current <= max) current else max
-    val percent = now.toFloat() / max
-    val progressBars = (totalBars * percent).toInt()
-    return (Strings.repeat("" + completedColor + symbol, progressBars)
-            + Strings.repeat("" + notCompletedColor + symbol, totalBars - progressBars))
 }
